backend/requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14
django-cors-headers>=4.3
python-dotenv>=1.0
supabase>=2.3
psycopg2-binary>=2.9
web3>=6.11
eth-account>=0.10
requests>=2.31
gunicorn>=21.2
whitenoise>=6.6
django-filter>=23.5
drf-spectacular>=0.27
celery>=5.3
redis>=5.0
scikit-learn>=1.3
pandas>=2.1
numpy>=1.26
joblib>=1.3
